ln_with_home first workings
I've created the ln_with_home script that links you can run to replaced
your home directory with links to the ./home folder.

I've fixed the home_mv_then_ln so that it appends to the ln_mounts.txt
file instead of over writting the file. I've also fixed the
ln_mounts.txt file which had an error caused by this problem.

I've tested using the ln_with_home on a new computer. However currently
it appears that vim minpac has not need added to the repo properly. I'll
need to investage this.